PWX-32885 & PWX-33218: Consolidate pre-flight conditional checks.
