+++ pkg/controller/storagecluster/controller_test.go
@@ -9058,262 +9058,6 @@ func TestDoesTelemetryMatch(t *testing.T) {
 	}
 }
 
-func TestShouldPreflightRun(t *testing.T) {
-	mockCtrl := gomock.NewController(t)
-	defer mockCtrl.Finish()
-
-	cluster := &corev1.StorageCluster{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      "px-cluster",
-			Namespace: "kube-test",
-		},
-	}
-
-	k8sVersion, _ := version.NewVersion(minSupportedK8sVersion)
-	driver := testutil.MockDriver(mockCtrl)
-	k8sClient := testutil.FakeK8sClient(cluster)
-	recorder := record.NewFakeRecorder(10)
-	controller := Controller{
-		client:            k8sClient,
-		Driver:            driver,
-		recorder:          recorder,
-		kubernetesVersion: k8sVersion,
-	}
-
-	// TestCase: aws cloud provider with image < 3.0
-	logrus.Infof("check aws cloud w/PX < 3.0...")
-	fakeK8sNodes := &v1.NodeList{Items: []v1.Node{
-		{ObjectMeta: metav1.ObjectMeta{Name: "node1"}, Spec: v1.NodeSpec{ProviderID: "aws://node-id-1"}},
-	}}
-	coreops.SetInstance(coreops.New(fakek8sclient.NewSimpleClientset(fakeK8sNodes)))
-
-	cluster.Spec.Image = "portworx/oci-image:2.9.0"
-
-	err := preflight.InitPreflightChecker(k8sClient)
-	require.NoError(t, err)
-
-	require.Equal(t, string(cloudops.AWS), pxutil.GetCloudProvider(cluster)) // Make sure aws
-
-	driver.EXPECT().UpdateDriver(gomock.Any())
-	driver.EXPECT().SetDefaultsOnStorageCluster(gomock.Any())
-	err = controller.setStorageClusterDefaults(cluster)
-	require.NoError(t, err)
-	require.False(t, preflightShouldRun(cluster))
-	logrus.Infof("aws cloud w/PX < 3.0, preflight will not run")
-
-	// Reset preflight for other tests
-	err = preflight.InitPreflightChecker(k8sClient)
-	require.NoError(t, err)
-
-	// TestCase: aws cloud provider with image >= 3.0
-	logrus.Infof("check aws cloud w/PX >= 3.0...")
-	cluster.Spec.Image = "portworx/oci-image:3.0.0"
-
-	err = preflight.InitPreflightChecker(k8sClient)
-	require.NoError(t, err)
-
-	driver.EXPECT().UpdateDriver(gomock.Any())
-	driver.EXPECT().SetDefaultsOnStorageCluster(gomock.Any())
-	err = controller.setStorageClusterDefaults(cluster)
-	require.NoError(t, err)
-	require.True(t, preflightShouldRun(cluster))
-	logrus.Infof("aws cloud w/PX >= 3.0, preflight will run")
-
-	// Reset preflight for other tests
-	coreops.SetInstance(coreops.New(fakek8sclient.NewSimpleClientset()))
-	err = preflight.InitPreflightChecker(k8sClient)
-	require.NoError(t, err)
-
-	// TestCase: Vsphere cloud provider with image <= 3.0
-	logrus.Infof("check vsphere cloud w/PX <= 3.0...")
-	cluster.Spec.Image = "portworx/oci-image:3.0.0"
-
-	// force vsphere
-	env := make([]v1.EnvVar, 2)
-	env[0].Name = "VSPHERE_VCENTER"
-	env[0].Value = "some.vcenter.server.com"
-	cluster.Spec.Env = env
-
-	require.Equal(t, string(cloudops.Vsphere), pxutil.GetCloudProvider(cluster)) // Make sure Vsphere
-
-	err = preflight.InitPreflightChecker(k8sClient)
-	require.NoError(t, err)
-
-	driver.EXPECT().UpdateDriver(gomock.Any())
-	driver.EXPECT().SetDefaultsOnStorageCluster(gomock.Any())
-	err = controller.setStorageClusterDefaults(cluster)
-	require.NoError(t, err)
-	require.False(t, preflightShouldRun(cluster))
-	logrus.Infof("vshpere cloud w/PX <= 3.0, preflight will not run")
-
-	// Reset preflight for other tests
-	err = preflight.InitPreflightChecker(k8sClient)
-	require.NoError(t, err)
-
-	// TestCase: Vsphere cloud provider with image >= 3.1
-	logrus.Infof("check vsphere cloud w/PX >= 3.1...")
-	cluster.Spec.Image = "portworx/oci-image:3.1.0"
-
-	err = preflight.InitPreflightChecker(k8sClient)
-	require.NoError(t, err)
-
-	driver.EXPECT().UpdateDriver(gomock.Any())
-	driver.EXPECT().SetDefaultsOnStorageCluster(gomock.Any())
-	err = controller.setStorageClusterDefaults(cluster)
-	require.NoError(t, err)
-	require.True(t, preflightShouldRun(cluster))
-	logrus.Infof("vshpere cloud w/PX >= 3.1, preflight will run")
-
-	// TestCase: Vsphere cloud provider with Install mode 'local'
-	logrus.Infof("check vsphere cloud w/local install mode...")
-	cluster.Spec.Image = "portworx/oci-image:3.1.0"
-	env[1].Name = "VSPHERE_INSTALL_MODE"
-	env[1].Value = "local"
-	cluster.Spec.Env = env
-
-	err = preflight.InitPreflightChecker(k8sClient)
-	require.NoError(t, err)
-
-	driver.EXPECT().UpdateDriver(gomock.Any())
-	driver.EXPECT().SetDefaultsOnStorageCluster(gomock.Any())
-	err = controller.setStorageClusterDefaults(cluster)
-	require.NoError(t, err)
-	require.False(t, preflightShouldRun(cluster))
-	logrus.Infof("vshpere cloud w/local install mode will not run")
-
-	// Reset preflight for other tests
-	coreops.SetInstance(coreops.New(fakek8sclient.NewSimpleClientset()))
-	err = preflight.InitPreflightChecker(k8sClient)
-	require.NoError(t, err)
-
-	// TestCase: Pure cloud provider with image <= 3.0
-	logrus.Infof("check Pure cloud w/PX <= 3.0...")
-	cluster.Spec.Image = "portworx/oci-image:3.0.0"
-
-	// force Pure
-	env = make([]v1.EnvVar, 1)
-	env[0].Name = "PURE_FLASHARRAY_SAN_TYPE"
-	env[0].Value = "FC"
-	cluster.Spec.Env = env
-
-	require.Equal(t, string(cloudops.Pure), pxutil.GetCloudProvider(cluster)) // Make sure Pure
-
-	err = preflight.InitPreflightChecker(k8sClient)
-	require.NoError(t, err)
-
-	driver.EXPECT().UpdateDriver(gomock.Any())
-	driver.EXPECT().SetDefaultsOnStorageCluster(gomock.Any())
-	err = controller.setStorageClusterDefaults(cluster)
-	require.NoError(t, err)
-	require.False(t, preflightShouldRun(cluster))
-	logrus.Infof("Pure cloud w/PX <= 3.0, preflight will not run")
-
-	// Reset preflight for other tests
-	err = preflight.InitPreflightChecker(k8sClient)
-	require.NoError(t, err)
-
-	// TestCase: Pure cloud provider with image >= 3.1
-	logrus.Infof("check Pure cloud w/PX >= 3.1...")
-	cluster.Spec.Image = "portworx/oci-image:3.1.0"
-
-	err = preflight.InitPreflightChecker(k8sClient)
-	require.NoError(t, err)
-
-	driver.EXPECT().UpdateDriver(gomock.Any())
-	driver.EXPECT().SetDefaultsOnStorageCluster(gomock.Any())
-	err = controller.setStorageClusterDefaults(cluster)
-	require.NoError(t, err)
-	require.True(t, preflightShouldRun(cluster))
-	logrus.Infof("Pure cloud w/PX >= 3.1, preflight will run")
-
-	// Reset preflight for other tests
-	coreops.SetInstance(coreops.New(fakek8sclient.NewSimpleClientset()))
-	err = preflight.InitPreflightChecker(k8sClient)
-	require.NoError(t, err)
-
-	// TestCase: Azure cloud provider with image <= 3.0
-	logrus.Infof("check Azure cloud w/PX <= 3.0...")
-	fakeK8sNodes = &v1.NodeList{Items: []v1.Node{
-		{ObjectMeta: metav1.ObjectMeta{Name: "node1"}, Spec: v1.NodeSpec{ProviderID: "azure://node-id-1"}},
-	}}
-
-	versionClient := fakek8sclient.NewSimpleClientset(fakeK8sNodes)
-	versionClient.Discovery().(*fakediscovery.FakeDiscovery).FakedServerVersion = &kversion.Info{
-		GitVersion: "v1.26.5",
-	}
-	coreops.SetInstance(coreops.New(versionClient))
-	k8sClient = testutil.FakeK8sClient(fakeK8sNodes)
-
-	cluster.Spec.Image = "portworx/oci-image:3.0.0"
-
-	err = preflight.InitPreflightChecker(k8sClient)
-	require.NoError(t, err)
-
-	c := preflight.Instance()
-	require.Equal(t, cloudops.Azure, c.ProviderName()) // Make sure Pure
-
-	driver.EXPECT().UpdateDriver(gomock.Any())
-	driver.EXPECT().SetDefaultsOnStorageCluster(gomock.Any())
-	err = controller.setStorageClusterDefaults(cluster)
-	require.NoError(t, err)
-	require.False(t, preflightShouldRun(cluster))
-	logrus.Infof("Azure cloud w/PX <= 3.0, preflight will not run")
-
-	// Reset preflight for other tests
-	err = preflight.InitPreflightChecker(k8sClient)
-	require.NoError(t, err)
-
-	// TestCase: Azure cloud provider with image >= 3.1
-	logrus.Infof("check Azure cloud w/PX >= 3.1...")
-	cluster.Spec.Image = "portworx/oci-image:3.1.0"
-
-	err = preflight.InitPreflightChecker(k8sClient)
-	require.NoError(t, err)
-
-	driver.EXPECT().UpdateDriver(gomock.Any())
-	driver.EXPECT().SetDefaultsOnStorageCluster(gomock.Any())
-	err = controller.setStorageClusterDefaults(cluster)
-	require.NoError(t, err)
-	require.True(t, preflightShouldRun(cluster))
-	logrus.Infof("Pure Azure w/PX >= 3.1, preflight will run")
-
-	// Reset preflight for other tests
-	coreops.SetInstance(coreops.New(fakek8sclient.NewSimpleClientset()))
-	err = preflight.InitPreflightChecker(k8sClient)
-	require.NoError(t, err)
-
-	// TestCase: Vsphere cloud provider with PKS and image >= 3.1
-	logrus.Infof("check vsphere cloud with PKS and PX >= 3.1...")
-	cluster.Spec.Image = "portworx/oci-image:3.1.0"
-
-	env = make([]v1.EnvVar, 1)
-	env[0].Name = "VSPHERE_VCENTER"
-	env[0].Value = "some.vcenter.server.com"
-	cluster.Spec.Env = env
-
-	coreops.SetInstance(coreops.New(fakek8sclient.NewSimpleClientset()))
-	k8sClient = testutil.FakeK8sClient(cluster)
-	ns := &v1.Namespace{
-		ObjectMeta: metav1.ObjectMeta{
-			Name: preflight.PksSystemNamespace,
-		},
-	}
-	_, err = coreops.Instance().CreateNamespace(ns)
-	require.NoError(t, err)
-
-	err = preflight.InitPreflightChecker(k8sClient)
-	require.NoError(t, err)
-	require.True(t, preflight.IsPKS())
-
-	driver.EXPECT().UpdateDriver(gomock.Any())
-	driver.EXPECT().SetDefaultsOnStorageCluster(gomock.Any())
-	err = controller.setStorageClusterDefaults(cluster)
-	require.NoError(t, err)
-	require.False(t, preflightShouldRun(cluster))
-	logrus.Infof("vsphere cloud with PKS and PX >= 3.1 will not run")
-
-}
-
 func TestEKSPreflightCheck(t *testing.T) {
 	mockCtrl := gomock.NewController(t)
 	defer mockCtrl.Finish()
