+++ pkg/controller/storagecluster/storagecluster.go
@@ -741,36 +741,6 @@ func (c *Controller) miscCleanUp(cluster *corev1.StorageCluster) error {
 	return nil
 }
 
-func preflightShouldRun(cluster *corev1.StorageCluster) bool {
-	pxVer30, _ := version.NewVersion("3.0")
-	pxVer31, _ := version.NewVersion("3.1")
-
-	// Preflight should only run freshInstall and if the PX version is 3.0.0 and above
-	if pxutil.IsFreshInstall(cluster) {
-		if pxutil.GetPortworxVersion(cluster).GreaterThanOrEqual(pxVer30) {
-			if !pxutil.IsVsphere(cluster) && !pxutil.IsPure(cluster) && !preflight.IsAzure() {
-				return true
-			}
-
-			if pxutil.IsVsphere(cluster) {
-				if preflight.IsPKS() { // Don't run Vsphere w/PKS
-					return false
-				}
-
-				envValue, exists := pxutil.GetClusterEnvValue(cluster, "VSPHERE_INSTALL_MODE")
-				if exists && envValue == pxutil.VsphereInstallModeLocal { // Don't run Vsphere w/local install mode
-					return false
-				}
-			}
-			// Vsphere, Pure & Azure only supported on 3.1.0
-			if pxutil.GetPortworxVersion(cluster).GreaterThanOrEqual(pxVer31) {
-				return true
-			}
-		}
-	}
-	return false
-}
-
 func (c *Controller) syncStorageCluster(
 	cluster *corev1.StorageCluster,
 ) error {
@@ -790,17 +760,9 @@ func (c *Controller) syncStorageCluster(
 			cluster.Namespace, cluster.Name, err)
 	}
 
-	if preflightShouldRun(cluster) {
-		// If preflight failed, or previous check failed, reconcile would stop here until issues got resolved
-		if err := c.runPreflightCheck(cluster); err != nil {
-			return fmt.Errorf("preflight check failed for StorageCluster %v/%v: %v", cluster.Namespace, cluster.Name, err)
-		}
-	} else {
-		// Always disable preflight if not supported.
-		if cluster.Annotations == nil {
-			cluster.Annotations = make(map[string]string)
-		}
-		cluster.Annotations[pxutil.AnnotationPreflightCheck] = "false"
+	// If preflight failed, or previous check failed, reconcile would stop here until issues got resolved
+	if err := c.runPreflightCheck(cluster); err != nil {
+		return fmt.Errorf("preflight check failed for StorageCluster %v/%v: %v", cluster.Namespace, cluster.Name, err)
 	}
 
 	if err := c.miscCleanUp(cluster); err != nil {
