+++ pkg/preflight/utils.go
@@ -24,6 +24,11 @@ func IsAzure() bool {
 	return Instance().ProviderName() == string(cloudops.Azure)
 }
 
+// IsAWS() returns whether the cloud environment is running on AWS
+func IsAWS() bool {
+	return Instance().ProviderName() == string(cloudops.AWS)
+}
+
 // RequiresCheck returns whether a preflight check is needed based on the platform
 func RequiresCheck() bool {
 	return Instance().ProviderName() == string(cloudops.AWS) ||
